requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
groq==0.4.2
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.24.1
requests>=2.31.0
typing-extensions>=4.8.0
streamlit>=1.28.0
